backend/requirements.txt
# 42Nexus Backend Requirements
# This file is for: ADMIRAL (Backend Dev 1) & ZERO (Backend Dev 2)

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
